# specs/009-password-reset/contracts/password-reset.openapi.yaml
openapi: 3.1.0
info:
  title: BePro Password Reset
  version: "1.0.0"
  description: |
    Self-service password reset for active users via tokenized email link.
    Both endpoints are public (no auth) and intentionally enumeration-safe:
    every accepted call returns the same shape regardless of whether the
    email or token corresponds to a real user.

    Implements feature 009. Spec: ../spec.md.

servers:
  - url: https://api.bepro.example/api
    description: Production
  - url: http://localhost:8787/api
    description: Local Wrangler dev

paths:
  /auth/password-reset/request:
    post:
      summary: Request a password-reset email
      description: |
        Always returns 200 with the same payload regardless of whether the
        email belongs to an active user (FR-001, User Story 2). Rate-limited
        per email (1/min, 5/hour) — throttled requests look identical to
        accepted ones (FR-009). On success for an active user, dispatches a
        Spanish-language email containing a 30-min-TTL link of the form
        `${APP_URL}/reset-password?token=...` and writes a
        `password_reset_requested` audit row (FR-006, FR-011).
      operationId: requestPasswordReset
      security: []  # public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestBody"
      responses:
        "200":
          description: Always returned. Generic success copy. No information leak.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetRequestResponse"
        "400":
          description: Malformed JSON or invalid email format. Distinct from "unknown email" (which is 200).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/password-reset/confirm:
    post:
      summary: Consume a reset token and set a new password
      description: |
        Validates the token (single-use, 30-min TTL), sets the new bcrypt
        hash (cost ≥ 12), revokes all active refresh tokens for that user,
        clears any active brute-force lockout, marks the token used, writes
        a `password_reset_completed` audit row, and issues a fresh session.
        Response shape is identical to `POST /auth/login` so the web client
        can re-use the post-login flow (FR-002, FR-012).

        On any token failure (expired, used, invalid, revoked because user
        deactivated, etc.) returns the same 400 with the same Spanish copy
        — no differentiation (FR-014, edge cases E-01/E-02/E-03).
      operationId: confirmPasswordReset
      security: []  # public — auth is the token itself
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmBody"
      responses:
        "200":
          description: Token valid; password set; session issued. Sets `refresh_token` httpOnly cookie like /auth/login.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "refresh_token=…; Path=/api/auth; HttpOnly; Secure; SameSite=Strict; Max-Age=604800"
        "400":
          description: Token expired, used, invalid, or user deactivated. Single generic message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "el enlace ha expirado, solicita uno nuevo"

components:
  schemas:
    PasswordResetRequestBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: "juan@ejemplo.com"

    PasswordResetRequestResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Generic Spanish confirmation. Same for valid, invalid, and rate-limited callers.
          example: "Si la cuenta existe, te hemos enviado un enlace para restablecer tu contraseña."

    PasswordResetConfirmBody:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          description: 32 bytes URL-safe base64 (no padding). 43 ASCII chars exactly.
          minLength: 43
          maxLength: 43
          pattern: "^[A-Za-z0-9_-]{43}$"
        password:
          type: string
          description: New password. ≥ 8 chars; at least one letter, one digit, one non-alphanumeric (FR-008).
          minLength: 8
          maxLength: 128

    AuthResult:
      type: object
      description: Same shape as `POST /auth/login` (FR-002).
      required: [accessToken, expiresAt, user]
      properties:
        accessToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/CurrentUser"

    CurrentUser:
      type: object
      required: [id, email, firstName, lastName, role, tenantId, isFreelancer, mustChangePassword]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        role:
          type: string
          enum: [admin, manager, account_executive, recruiter]
        tenantId: { type: string, format: uuid }
        isFreelancer: { type: boolean }
        mustChangePassword:
          type: boolean
          description: Always `false` after a successful reset (the user just chose this password themselves).

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

  securitySchemes: {}